+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Changed license version from GPL_v3 to dual MIT & APACHE-2.0 (https://github.com/nymtech/nym-bridges/pull/63)
+- attempt connection open to all addresses provided in a ClientConfig using happy eyeballs (https://github.com/nymtech/nym-bridges/pull/64)
+- removed filter to ipv4 only from quic_plain transport (https://github.com/nymtech/nym-bridges/pull/64)
 
 ### Fixed